% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidyfood_update.R
\name{check_tidyfood_version}
\alias{check_tidyfood_version}
\title{check_tidyfood_version}
\usage{
check_tidyfood_version(
  packages = c("core", "all"),
  from = c("gitlab", "github", "gitee", "shen")
)
}
\arguments{
\item{packages}{core or all packages in tidyfood. "core" means all the core
packages in tidyfood and "all" means all the packages in tidyfood.}

\item{from}{"gitlab", "github", or "gitee", or "tidyfood.org"}
}
\description{
Check if there are packages in tidyfood can be updated.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
