% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_tidyfood_logo.R
\name{tidyfood_logo}
\alias{tidyfood_logo}
\title{Show the logo tidyfood}
\usage{
tidyfood_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\value{
A ASCII log of tidyfood
}
\description{
The tidyfood logo, using ASCII or Unicode characters
}
\examples{
tidyfood_logo()
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
