% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_utils.R
\name{tidyfood_packages}
\alias{tidyfood_packages}
\title{List all packages in the tidyfood}
\usage{
tidyfood_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidyfood in the list?}
}
\description{
List all packages in the tidyfood
}
\examples{
tidyfood_packages()
}
