% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_tidyfood_update.R
\name{update_tidyfood}
\alias{update_tidyfood}
\title{update_tidyfood}
\usage{
update_tidyfood(
  packages = c("core", "all"),
  from = c("gitlab", "github", "gitee", "tidyfood.org"),
  fastgit = FALSE
)
}
\arguments{
\item{packages}{core or all packages in tidyfood. "core" means all the core
packages in tidyfood and "all" means all the packages in tidyfood.}

\item{from}{github, gitlab or gitee.}

\item{fastgit}{if install packages using fastgit. see
https://hub.fastgit.org/}
}
\description{
Update packages in tidyfood.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
